requirements.txt
transformers
torch
sentencepiece
sacremoses
Pillow
einops
